automatically backed up by learn

+++ lib/bachelor.rb
@@ -52,7 +52,7 @@ def get_average_age_for_season(data, season)
     if k == season
       v.each do |person|
         count += 1 
-        out += person[:age].to_i
+        out += person[:age].to_f
       end
     end
   end
